+++ src/components/index.js
@@ -1,6 +1,7 @@
 export { default as CustomSwitch } from './CustomSwitch';
 export { default as CustomTooltip } from './CustomTooltip';
 export { default as CustomInput } from './CustomInput';
+export { default as CustomInputPassword } from './CustomInputPassword';
 export { default as Spinner } from './Spinner';
 export { default as Navigation } from './Navigation';
 export { default as GeneralError } from './GeneralError';