+++ src/views/index.js
@@ -1,4 +1,5 @@
-export { default as Home } from './Home';
 export { default as ValidateCredentials } from './ValidateCredentials';
+export { default as Home } from './Home';
+export { default as SignInAdmin } from './SignInAdmin';
 export { default as Error401 } from './Errors/401';
 export { default as Error404 } from './Errors/404';